Pulled logic out of OrderDetailsController
