+++ src/Foundation/Features/MyAccount/OrderDetails/OrderDetailsController.cs
@@ -1,13 +1,7 @@
-﻿using EPiServer;
-using EPiServer.Commerce.Order;
-using EPiServer.Logging;
+﻿using EPiServer.Commerce.Order;
 using EPiServer.Web.Mvc;
-using Foundation.Commerce;
-using Foundation.Commerce.Customer.Services;
-using Foundation.Commerce.Customer.ViewModels;
 using Foundation.Commerce.Models.Pages;
 using Foundation.Commerce.Order.Services;
-using Foundation.Commerce.Order.ViewModels;
 using Mediachase.Commerce.Orders;
 using System;
 using System.Collections.Generic;
@@ -19,27 +13,21 @@ namespace Foundation.Features.MyAccount.OrderDetails
 {
     public class OrderDetailsController : PageController<OrderDetailsPage>
     {
-        private readonly IAddressBookService _addressBookService;
         private readonly IOrdersService _ordersService;
-        private readonly ICustomerService _customerService;
         private readonly IOrderRepository _orderRepository;
-        private readonly IContentLoader _contentLoader;
-        private readonly ICartService _cartService;
-        private readonly IPurchaseOrderFactory _purchaseOrderFactory;
+        private readonly OrderDetailsControllerService _controllerService;
 
-        public OrderDetailsController(IAddressBookService addressBookService, IOrdersService ordersService, ICustomerService customerService, IOrderRepository orderRepository, IContentLoader contentLoader, ICartService cartService, IPurchaseOrderFactory purchaseOrderFactory)
+        public OrderDetailsController(IOrdersService ordersService,
+            IOrderRepository orderRepository,
+            OrderDetailsControllerService controllerService)
         {
-            _addressBookService = addressBookService;
             _ordersService = ordersService;
-            _customerService = customerService;
             _orderRepository = orderRepository;
-            _contentLoader = contentLoader;
-            _cartService = cartService;
-            _purchaseOrderFactory = purchaseOrderFactory;
+            _controllerService = controllerService ?? throw new ArgumentNullException(nameof(controllerService));
         }
 
         [HttpGet]
-        public ActionResult Index(OrderDetailsPage currentPage, int orderGroupId = 0) => View(GetModel(orderGroupId, currentPage));
+        public ActionResult Index(OrderDetailsPage currentPage, int orderGroupId = 0) => View(_controllerService.GetViewModel(orderGroupId, currentPage));
 
         [HttpPost]
         [ValidateAntiForgeryToken]
@@ -71,7 +59,7 @@ namespace Foundation.Features.MyAccount.OrderDetails
         public ActionResult ChangePrice(int orderGroupId, int shipmentId, int lineItemId, decimal placedPrice, OrderDetailsPage currentPage)
         {
             var issues = _ordersService.ChangeLineItemPrice(orderGroupId, shipmentId, lineItemId, placedPrice);
-            var model = GetModel(orderGroupId, currentPage);
+            var model = _controllerService.GetViewModel(orderGroupId, currentPage);
             model.ErrorMessage = GetValidationMessages(issues);
             return PartialView("Index", model);
         }
@@ -80,7 +68,7 @@ namespace Foundation.Features.MyAccount.OrderDetails
         public ActionResult ChangeQuantity(int orderGroupId, int shipmentId, int lineItemId, decimal quantity, OrderDetailsPage currentPage)
         {
             var issues = _ordersService.ChangeLineItemQuantity(orderGroupId, shipmentId, lineItemId, quantity);
-            var model = GetModel(orderGroupId, currentPage);
+            var model = _controllerService.GetViewModel(orderGroupId, currentPage);
             model.ErrorMessage = GetValidationMessages(issues);
             return PartialView("Index", model);
         }
@@ -107,79 +95,6 @@ namespace Foundation.Features.MyAccount.OrderDetails
             }
 
             return string.Join(".", messages);
-        }
-
-        private OrderDetailsViewModel GetModel(int orderGroupId, OrderDetailsPage currentPage)
-        {
-            var orderViewModel = new OrderDetailsViewModel
-            {
-                CurrentContent = currentPage,
-                CurrentCustomer = _customerService.GetCurrentContactViewModel()
-            };
-
-            var purchaseOrder = OrderContext.Current.Get<PurchaseOrder>(orderGroupId);
-            if (purchaseOrder == null)
-            {
-                return orderViewModel;
-            }
-
-            // Assume there is only one form per purchase.
-            var form = purchaseOrder.GetFirstForm();
-
-            var billingAddress = form.Payments.FirstOrDefault() != null
-                ? form.Payments.First().BillingAddress
-                : new OrderAddress();
-
-            orderViewModel.PurchaseOrder = purchaseOrder;
-
-            orderViewModel.Items = form.Shipments.SelectMany(shipment => shipment.LineItems.Select(lineitem => new OrderDetailsItemViewModel
-            {
-                LineItem = lineitem,
-                Shipment = shipment,
-                PurchaseOrder = orderViewModel.PurchaseOrder as PurchaseOrder,
-            }
-            ));
-
-            orderViewModel.BillingAddress = _addressBookService.ConvertToModel(billingAddress);
-            orderViewModel.ShippingAddresses = new List<AddressModel>();
-
-            foreach (var orderAddress in form.Shipments.Select(s => s.ShippingAddress))
-            {
-                var shippingAddress = _addressBookService.ConvertToModel(orderAddress);
-                orderViewModel.ShippingAddresses.Add(shippingAddress);
-                orderViewModel.OrderGroupId = purchaseOrder.OrderGroupId;
-            }
-            if (purchaseOrder[Constant.Quote.QuoteExpireDate] != null &&
-                !string.IsNullOrEmpty(purchaseOrder[Constant.Quote.QuoteExpireDate].ToString()))
-            {
-                DateTime.TryParse(purchaseOrder[Constant.Quote.QuoteExpireDate].ToString(), out var quoteExpireDate);
-                if (DateTime.Compare(DateTime.Now, quoteExpireDate) > 0)
-                {
-                    orderViewModel.QuoteStatus = Constant.Quote.QuoteExpired;
-                    try
-                    {
-                        // Update order quote status to expired
-                        purchaseOrder[Constant.Quote.QuoteStatus] = Constant.Quote.QuoteExpired;
-                        _orderRepository.Save(purchaseOrder);
-                    }
-                    catch (Exception ex)
-                    {
-                        LogManager.GetLogger(GetType()).Error("Failed to update order status to Quote Expired.", ex.StackTrace);
-                    }
-
-                }
-            }
-
-
-            if (!string.IsNullOrEmpty(purchaseOrder["QuoteStatus"]?.ToString()) &&
-                (purchaseOrder.Status == OrderStatus.InProgress.ToString() ||
-                 purchaseOrder.Status == OrderStatus.OnHold.ToString()))
-            {
-                orderViewModel.QuoteStatus = purchaseOrder["QuoteStatus"].ToString();
-            }
-
-            orderViewModel.BudgetPayment = _ordersService.GetOrderBudgetPayment(purchaseOrder);
-            return orderViewModel;
         }
     }
 }
